+++ index.js
@@ -1,11 +1,11 @@
 const {glob}=require('glob')
 const cron=require('node-cron')
-const {writeFile}=require('fs')
+const {readFileSync,writeFile}=require('fs')
 const {basename}=require('path')
 const {Client}=require('discord.js')
-const client=new Client()
-const config=require('./config.json')
-let database=require('./database.json')
+global.client=new Client()
+global.config=require('./config.json')
+global.database=JSON.parse(readFileSync('database.json'))
 let commands=[]
 glob('commands/**/*.js',(err,files)=>{
 	if(err)console.log(err)
@@ -22,6 +22,7 @@ client.on('guildMemberAdd',async member=>{
 	if(!member.user.bot&&!database['users'][member.id]){
 		database['users'][member.id]={
 			warns : 0,
+			money : 100
 		}
 	}
 })
@@ -38,18 +39,7 @@ client.on('message',async message=>{
 		const args=content.slice(1)
 		message.delete()
 		if(command in commands){
-			commands[command].run(client,database,config,message,args)
-		}else{
-			if(!message.member.id==config.ownerID)
-				message.channel.send('```Nie możesz użyć tej komendy.```')
-			switch(command){
-				case 'dbsave':
-					writeFile('database.json',JSON.stringify(database,null,4),function(){})
-					break;
-				case 'dbload':
-					database=require('./database.json')
-					break;
-			}
+			commands[command].run(message,args)
 		}
 	}
 })